Update patch set 3
Patch Set 3:

The change seems fine, although the cleaner fix would be so that the check for whether the current frame will be coded intra should move to before calling "vp9_get_compressed_data" itself (or at the beginning of the function) to avoid re-calling many functions unnecessarily. The change is good for now, but we should clean this up later.

Patch-set: 3