Update patch set 6
Patch Set 6:

Hi Yunqing. Please do check, I might have just missed somethig. Regarding vp10_setup_past_independence, the difference on nextgen between the 2 versions of this functions is that the _dec_ version only memsets the necessary part of cm->mip (that is if only a particular tile is decoded). On nextenv2, this memset was not there in the first place and some git archaelogy suggested to me that it was never there in nextgenv2, and in nextgen it was added significantly befre vp10 was forked, so I assumed it is not necessary. Please let me know otherwise.

Patch-set: 6