Update patch set 6
Patch Set 6:

There were a lot of conflicts as the original tile coding on nextgenv2 has been changed compared to nextgen. There was also a lot of replication of functionality between the baseline and the ext-tile code on nextgenv2 so I coalesced a lot of these. I have tested encoding performance using a single tile before and after the patch and there was no degradation in this single tile use case (there was a little variation due to the uncompressed frame header signalling the tile sizes differently, but this was in the noise in my tests).  Are you using the same tile sizes? On nextgenv2 the tile-columns/tile-rows (width/height) are 1 based, while on nextgen they were 0 based I believe. Can you tell me how to reproduce the decoder crash?

Patch-set: 6