Update patch set 1
Patch Set 1:

In the CQ mode, the concept is that there is no constraint on how low the bit-rate can go as long as the minimum quality defined by the cq-level parameter is met. So for mid to hard content, it should behave exactly like vbr mode, only for easier content the behavior should become more like q. The only problem is knowing what the minimum quality level is for easier content, and how it relates to the q-index that is used as the lower bound. The only thing this patch does is to use a perceptual consideration to redefine this mapping between the cq-level keyed in and the minimum qindex used. The adjustment is made only at the smoother end because I think that is where I think perceptual effects and qindex diverge more. But this mapping could evolve over time.

The approach you suggest is more generic and could/should be applied for either vbr or cq. And if we do it for vbr, it will automatically be used for cq as well. Perhaps we should introduce a new parameter like --psy-rd to change how the bit allocation is made.

Patch-set: 1
Label: Code-Review=0