Update patch set 2
Patch Set 2:

> Sorry for the confusion on the review.
 > 
 > If the issue is as what has been addressed in the comments, I would
 > suggest to change the settings in
 > 
 > static int get_ref_frame_flags(const VP10_COMP *cpi)
 > 
 > in encoder.c, in a way such that the altref bit keeps set but to
 > remove the bit set for LAST{2,3,4}, whichever is the same as
 > altref, hence the bit setting for altref in cpi->ref_frame_flags
 > would be able to then set in the same way as when ext-ref is not
 > enabled.

Ok, we can go that way, but note that this will change the encoder output (for all encodes, not just the cpu-used >= 3 afected by the bug), as it will now code the ALTREF frame (see new commit mesage) as opposed to LAST* if they are the same (after key frames). This will affect both the baseline and the ext-refs experimant, so you might want to run a coding performance test to make sure the effect is in the noise. Patch set 2 does what you say, so you can go ahead.

The original patch produced the same encoder output for the cases that were not crashing, so there is no worry there.

Patch-set: 2