Update patch set 5
Patch Set 5:

If the input level is undefined, the encoder will exit with VPX_CODEC_INVALID_PARAM. I don't think we need test for invalid input parameters like this.

The other case is the input level is not achievable, e.g. targeting for level 1.0 with a 4K input video. We expect the encoder to handle such situations, and should add such test cases. I will do it when the corresponding encoder functionality is implemented in future patches.

Patch-set: 5
Label: Code-Review=0