Update patch set 1
Patch Set 1:

> In the future let's wait a day or so until we revert - since this
 > can cause more problems than in solves, given that we do not have
 > per commit coverage for the ext-inter experiment.

That's the opposite of my view. To me it's better to restore the tree to a working state to avoid new changes potentially compounding the issue or adding new ones. If a change is small and split on logical bounds it should revert cleanly especially in this short time frame.

Patch-set: 1