Update patch set 4
Patch Set 4:

was a bit hasty on the +2 as we were still discussing things offline. this should be fine once the additional function call is added to the test, but we should file a bug to sort out how the decrypt callback is being used in this function (i.e., test/vp9_decrypt_test.cc didn't fail with the smaller clear buffer).

Patch-set: 4
Label: Code-Review=0