Update patch set 5
Patch Set 5:

(10 comments)

Thanks for the comments!
I haven't fixed them all, but I want to push the new patch for your review.
In this patch several bugs are fixed. Other than the bug you mentioned, the other huge bug is that the upsampled reference frames are not updated correctly.

Patch-set: 5
Label: Code-Review=0