Update patch set 2
Patch Set 2:

The goal of all of this is to be able to merge aom and nextgenv2 as soon as possible, I hope we get the codebase to be similar to current aom code, to make a merge possible without incurring too many bugs. After that, we can further improve naming and so on.

Patch-set: 2
Reviewer: Yaowu Xu <1000893@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0