Update patch set 3
Patch Set 3: Code-Review-1

Unless there is a strong reason to re-apply this patch, I think we should not do so.

I do not have any benchmark data to support my opinion (not sure there are any relevant benchmarks for this code), but the way to fix TSan warnings proposed by this patch is suboptimal. It introduces per-row mutexes that are supposed to serve as atomic variables, but are actually slower (the original patch claimed that "The decoder speed tests showed < 3% speed loss while using 2 ~ 4 threads.").

IIRC, when we've discussed the revert of 360630, the consensus was that we'd better introduce proper atomic operations layer into vp8 and use it to synchronize the encoder. I'm still ready to help with that, if needed.

Patch-set: 3
Reviewer: Alexander Potapenko <1000853@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1