Update patch set 5
Patch Set 5:

As per the discussion over hangouts:  

pthread_barrier would be fine except for the fact that we'd have to write a version that ran on windows. This code does the same thing by creating a semaphore for each thread on exit and waiting on that.

Patch-set: 5
Reviewer: James Bankoski <1000963@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0