Update patch set 1
Patch Set 1:

my working theory is that my desktop needed to be rebooted (updated packages) and that was messing up the compiler somehow.

Sorry, that's the best I've got. Retested with arm gcc 4.8 on ubuntu and aarch64 gcc and clang from ndk r12b and qemu and a nexus 9.

Patch-set: 1