Update patch set 4
Patch Set 4:

It's not ideal but it seems like a good start. It is much faster than the 1024 neon implementation. Managed to get the intermediate buffer down to 8x32 (because that is the only output of the first pass) and wedge the 'add' portion into the second pass which removed the other buffer entirely.

I may be doing a transpose to the dest data that is not necessary. Need to fix up the test to make sure.

Patch-set: 4