Update patch set 1
Patch Set 1:

> I see the assert is missing from the aom branch. I kind of like
 > these actually using a prob that will trigger an assert in the
 > coder to keep them from sneaking into the bitstream accidentally,
 > but I suppose we can update UNUSED_PROB after making cost
 > initialization more conservative.

Yes, I know what you mean.
Unfortunately, couldn't think of an easy way to make cost initialization less conservative (as it doesn't know which probabilities will be actually used later).
So, as you said, we can go ahead with this for now, and if we find a solution to the assert, we can change the value of UNUSED_PROB.

Patch-set: 1
Label: Code-Review=0