Update patch set 1
Patch Set 1: Code-Review+2

Is the IF indent required? Seems to be in line with C defines to have them all the way to the left.

Might be interesting to have an #ifdef guard in vpx_config.asm but I'm not sure there is an #ifdef equivalent and this seems more than sufficient.

Patch-set: 1
Label: Code-Review=+2