Update patch set 6
Patch Set 6:

(8 comments)

> NEON/VpxPostProcDownAndAcrossMbRowTest.CheckCvsAssembly/0
 > test_libvpx.armv7-linux-gcc.gcc..rev2: vpx_dsp/arm/deblock_neon.c:95:
 > vpx_post_proc_down_and_across_mb_row_neon: Assertion `(size == 8 &&
 > cols % 8 == 0) || (size == 16 && cols % 16 == 0)' failed.

Yeah I explicitly made the function use %8 for the width to make sure it caught that case. Derestricted the condition since it'll handle %8 no problem. The MIPS on the other hand has separate functions for Y and U/V but it might be OK?

Patch-set: 6
Label: Code-Review=0