Update patch set 1
Patch Set 1:

This approach *may* be workable for the across filter. It requires holding a lot of things in registers though. Here we can load and discard below_context, but in the across filter it would need to transpose 8 at a time, making that difficult. (Basically need to keep 8x24 in registers - 8 left context, 8 source, 8 right context)

Patch-set: 1
Label: Code-Review=0