Update patch set 1
Patch Set 1:

Thank you for the detailed information. It makes much sense now. Except for the last frame in the video, I normally see rc->baseline_gf_interval > 1. Is it easy to be reproduced? Could you file an issue so that other people who is familiar with this code can also look at it?

Patch-set: 1