Update patch set 13
Patch Set 13:

> James/Jinging, [-]32640 was originally added in https://chromium-review.googlesource.com/313453
 > 
 > Any reason for that value over INT16_MIN/MAX?

Satd takes the coeff as input and the range of coeff is [-32640, 32640]. This is explained in the function vpx_hadamard_8x8_c and vpx_hadamard_16x16_c. Thus I think we should also use this value in the unit tests for vpx_block_error_fp. I'll change it back.

Patch-set: 13
Label: Code-Review=0