Update patch set 3
Patch Set 3:

The >10% here was in compression. In a clip like this the intra blocks coded in the sky (or even those coded inter but the intra score was also very low and close) were causing a suppression of the ARF boost.

Patch-set: 3
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>