Update patch set 5
Patch Set 5:

FYI: https://chromium-review.googlesource.com/c/306190/

Have you profiled this on a real encoder run or just the unit tests? Profiling speculative arithmetic optimizations on unit tests which squirt in random data is pointless. The job of the predictor in the real encoder is to make the residuals small, so the special case in the AVX code practically never kicks in. By all means if you want to throw away the AVX then do so, but then you should port the speculative arithmetic to the SSE2 code. On my tests for the original patch this optimization gave you about 1% total encoder speedup (VP9) on a Sandy Bridge server.

Patch-set: 5