Update patch set 4
Patch Set 4:

> Patch Set 1:
> 
> > Build Failed
>  > 
>  > https://build.webmproject.org/jenkins/job/libvpx__commit_trigger/34553/
>  > : FAILURE
> 
> It doesn't seem to like my @chromium.org address.
> Do I need to use my @google.com address to fix this?

It checks if you have previously uploaded anything *or* the uploader has an @google address. Typically the way around this is someone on the team rebases the change for you which changes the uploader but not the author. I didn't notice that you had already changed it to your @google.

Patch-set: 4
Reviewer: Johann Koenig <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>