Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > > Patch Set 2:
> > > > > 
> > > > > did you investigate the memcpy solution?
> > > > 
> > > > https://chromium-review.googlesource.com/c/503512/ uses memcpy as an alternative
> > > 
> > > right, I was wondering if there was a perf difference visible between the two. or if memcpy was creating OK code for the constant size
> > 
> > There is a perf difference. memcpy is not nearly as good as vst. It's a hack, not a useful solution
> 
> what does it generate?

I don't know exactly but it needs to allocate stack space, do the stores, and still make a copy. unaligned loads and stores are just as fast as aligned, so this is a whole lot of extra work.

> your loads should land first then and this rebased on that to pick up the gcc test

Yes, the other change does not have any parents. Waiting for it to be approved.

Patch-set: 2
Reviewer: Johann Koenig <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>