Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > what toolchain are you using to build this? does qemu have support for this, if so what version? related to qemu are there any android emulators with support?
> 
> I build and test libvpx on loongson machine, toolchain is gcc. qemu can emulate loongson platform, but libvpx havn't tested on it yet. If test machines are needed, please let me know the address to send to, my company will be willing to provide it.

The compiles that are done as part of the commit check are done with cross compilers under linux. The same is done with qemu for nightly tests as the infrastructure is hosted in google cloud platform. I don't know that hardware is strictly necessary at this time, we're just looking for general correctness on platforms that aren't directly supported by libvpx developers.

Patch-set: 1