Update patch set 15
Patch Set 15:

> Patch Set 14:
> 
> (2 comments)
> 
> Looks good, but I'm curious what's causing this change. Are you seeing large overshoots also for camera feeds?
> Also, aren't there any unit tests for these files?

The only scenario i've seen for camera feeds is when the camera is muted and then unmuted, then large overshoot can occur if the encoder is not reset when unmuted. WebRTC resets the VP8 encoder on mute/unmute, but they would like to remove the reset. The fix for VP9 has already been rolled in (for vp9, the encoder was not being reset). See issue:
https://bugs.chromium.org/p/webrtc/issues/detail?id=6213

Yes there are unittests in (e.g., test/datarate_test.cc) that cover the encoding path that uses this function/feature.

Patch-set: 15
Reviewer: Marco Paniconi <1126910@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>