Update patch set 5
Patch Set 5: Code-Review+2

(1 comment)

I am OK with this if VP8 is now mainly used for RTC. Even in other cases the compression improvement from the limited calls to full search is not that great most of the time and we are much less concerned now with trying to eek the last fraction of a % for best quality out of VP8. The one place where full search can sometimes help quite a lot is some types of animation and screen share content. It certainly results in the removal of a lot of code which is no bad thing.

Patch-set: 5
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2