Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > The Power instructions do not experience the same alignment penalty as Neon.
> 
> Cool, that sounds like a better way to fix this issue then.

Thanks for the suggestion. BUT we still have problem with "double".
Below is the problematic code that uses vsx (vector-scale-extension) builtin:

const int16x8_t l = unpack_to_s16_h(vec_vsx_ld(0, left));

The vec_vsx_ld here is translated into "loading [left,left+16) into a 128-bit vector register". The memory content is guaranteed to be in a vector register, so later unpack_to_s16_h could work on this vector register.

If we replaced this to read from "(double *)(left)", in the assembly code, the content could be in a double float register, or even a temporary memory location, anyway, very unlikely, if not possible, the compiler would load the content into the lower half of a vsx register (which is 128-bit). Also unpack_to_s16_h does not work on a double.

So if we were to use double for temporary results, we would end up rewriting all these vsx-builtins and the result would be a non-optimized version.

Patch-set: 8
Reviewer: Han Shen <1001115@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>