Update patch set 6
Patch Set 6:

(1 comment)

When I was working on the temporal filter it was passing unit tests but failing more generally. It turned out that the filter only applies if the difference between the frames is small, so I narrowed it to 0-7 and the test started failing so I could figure out what was up and fix it.

Patch-set: 6
Reviewer: Johann Koenig <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>